requirements.txt
# ========================================
# NVD Mirror - Python Dependencies
# ========================================
# Install with: pip install -r requirements.txt

# HTTP requests library
requests>=2.31.0

# PostgreSQL adapter
psycopg2-binary>=2.9.9

# Environment variable management
python-dotenv>=1.0.0
